% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cleanVarNames}
\alias{cleanVarNames}
\title{Clean variable names to a consistent format}
\usage{
cleanVarNames(dt)
}
\arguments{
\item{dt}{A data.frame or data.table.}
}
\value{
The modified \code{dt} with cleaned column names.
}
\description{
Replace spaces, slashes and punctuation commonly found in column names with
underscores and remove parentheses.
}
\examples{
cleanVarNames(df)
}
