% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getDataTimePeriod}
\alias{getDataTimePeriod}
\title{Wrapper to subset by a two-element time period}
\usage{
getDataTimePeriod(dt, dt_date_var, time_period)
}
\arguments{
\item{dt}{A data.table containing event records.}

\item{dt_date_var}{A string naming the date column in \code{dt} to filter on.}

\item{time_period}{A string naming an object (vector of length 2) containing start and end dates.}
}
\value{
A data.table filtered to the specified time period.
}
\description{
Calls \code{subsetEvents} using a named time period (vector of length 2).
}
\examples{
\dontrun{
getDataTimePeriod(events_dt, "event_date", "time_period_q1")
}
}
