% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{isLengthNonZero}
\alias{isLengthNonZero}
\title{Binary indicator aggregator for dcast}
\usage{
isLengthNonZero(x)
}
\arguments{
\item{x}{A vector passed by the aggregator.}
}
\value{
Integer 1 or 0.
}
\description{
Returns 1 when the length of x is greater than zero, otherwise 0. Intended
for use as \code{fun.aggregate} in \code{data.table::dcast} to create flags.
}
\examples{
isLengthNonZero(1:3)
}
